Add project pages

+++ src/hooks/useEntities.ts
@@ -1,19 +1,24 @@
 import useSWR from 'swr';
 
 export interface EntitiesHook<E> {
-  data: E[];
+  data: ReadonlyArray<Readonly<E>>;
   isLoading: boolean;
 
   save(entity: Omit<E, 'id'>): Promise<void>;
   deleteAll(): Promise<void>;
 }
 
-export default function useEntities<E>(url: string): Readonly<EntitiesHook<E>> {
+export default function useEntities<E>(
+  url: string | null,
+): Readonly<EntitiesHook<E>> {
   const { data, mutate, error, revalidate } = useSWR<E[], unknown>(url);
   return {
     data: data ?? [],
     isLoading: !error && !data,
     async save(entity: E) {
+      if (!url) {
+        throw new Error('Entities not ready');
+      }
       const tmpEntity: E = {
         ...entity,
         id: Math.random().toString(),
@@ -30,6 +35,9 @@ export default function useEntities<E>(url: string): Readonly<EntitiesHook<E>> {
       }
     },
     async deleteAll() {
+      if (!url) {
+        throw new Error('Entities not ready');
+      }
       await mutate([], false);
       const { ok } = await fetch(url, {
         method: 'DELETE',
